requirements.txt
pycurl
python-dotenv
bs4